--- pkg/network/tracer/testutil/gotls_client/.gitignore
@@ -1,1 +1,0 @@
-gotls_client
